+++ packages/flame/CHANGELOG.md
@@ -10,6 +10,7 @@
  - Rename `Tapable` to `Tappable`
  - Rename `HasTapableComponents` to `HasTappableComponents`
  - Add `intensity` to `Camera.shake`
+ - `FixedResolutionViewport` to use matrix transformations for `Canvas`
 
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories